install wxWidgets in CI

+++ debugger/data/Instruction.h
@@ -2,7 +2,7 @@
 #define RED_O_LATOR_INSTRUCTION_H
 
 struct Instruction {
-    size_t address;
+    uint64_t address;
     std::string text;
     bool isLabel = false;
 };
