+++ debugger/gui/hex-view/AddressRows.h
@@ -2,19 +2,21 @@
 #define RED_O_LATOR_ADDRESSROWS_H
 
 #include "../wx.h"
+#include <cstddef>
+#include <cstdint>
 
 class AddressRows : public wxScrolledWindow {
    public:
     explicit AddressRows(wxWindow* parent);
 
-    void setState(size_t address, size_t lineCount);
+    void setState(uint64_t address, size_t lineCount);
 
    private:
     size_t lineCount = 0;
     int lineHeight = 0;
     int lineWidth = 0;
 
-    size_t startAddress = 0;
+    uint64_t startAddress = 0;
 
     void OnDraw(wxDC& dc) override;
 };
